+++ app/Http/Controllers/Api/Partner/V1/CustomerController.php
@@ -2,11 +2,11 @@
 
 namespace App\Http\Controllers\Api\Partner\V1;
 
-use App\Dtos\CompanyDto;
+use App\Dtos\UserDto;
 use App\Http\Controllers\Controller;
-use App\JsonApi\Partner\V1\Companies\CompanyQuery;
-use App\JsonApi\Partner\V1\Companies\CompanyRequest;
-use App\JsonApi\Partner\V1\Companies\CompanySchema;
+use App\JsonApi\Partner\V1\Customers\CustomerQuery;
+use App\JsonApi\Partner\V1\Customers\CustomerRequest;
+use App\JsonApi\Partner\V1\Customers\CustomerSchema;
 use App\Models\Company;
 use App\Services\CompanyService;
 use App\JsonApi\Proxies\CompanyPartner; // proxy model https://laraveljsonapi.io/docs/1.0/digging-deeper/proxies.html
@@ -15,7 +15,7 @@ use LaravelJsonApi\Laravel\Http\Controllers\Actions;
 use LaravelJsonApi\Core\Document\Error;
 use LaravelJsonApi\Core\Responses\ErrorResponse;
 
-class CompanyController extends Controller
+class CustomerController extends Controller
 {
 
     use Actions\FetchMany;
@@ -29,16 +29,4 @@ class CompanyController extends Controller
     use Actions\AttachRelationship;
     use Actions\DetachRelationship;*/
 
-
-    /**
-     * Company service for creating and updating company
-     * @var CompanyService
-     */
-    protected CompanyService $companyService;
-
-
-    public function __construct()
-    {
-        $this->companyService = new CompanyService();
-    }
 }