+++ app/JsonApi/Partner/V1/Server.php
@@ -33,10 +33,11 @@ class Server extends BaseServer
     {
         return [
             // @TODO
-            Companies\CompanySchema::class,
+            Companies\PartnerSchema::class,
             Employees\EmployeeSchema::class,
             Account\AccountSchema::class,
             MyCompanies\MyCompanySchema::class,
+            Customers\CustomerSchema::class,
         ];
     }
 }