+++ app/Policies/Customer/MyCompanyPolicy.php
@@ -1,12 +1,13 @@
 <?php
 
-namespace App\Policies\Partner;
+namespace App\Policies\Customer;
 
-use App\JsonApi\Proxies\CompanyPartner as Company;
+//use App\Models\Partner;
+use App\JsonApi\Proxies\MyCompanyCustomer as Company;
 use App\Models\User;
 use Illuminate\Auth\Access\HandlesAuthorization;
 
-class CompanyPolicy
+class MyCompanyPolicy
 {
     use HandlesAuthorization;
 
@@ -18,20 +19,19 @@ class CompanyPolicy
      */
     public function viewAny(User $user)
     {
-        return true;
-       // return $user->isA('partner');
+        return $user->isA('customer');
     }
 
     /**
      * Determine whether the user can view the model.
      *
-     * @param  \App\Models\User  $user
-     * @param  \App\JsonApi\Proxies\CompanyPartner  $company
+     * @param \App\Models\User $user
+     * @param Company $company
      * @return \Illuminate\Auth\Access\Response|bool
      */
     public function view(User $user, Company $company)
     {
-        return $user->isA('partner');
+        return $user->isA('customer');
     }
 
     /**
@@ -42,46 +42,47 @@ class CompanyPolicy
      */
     public function create(User $user)
     {
-        return false;
-        // return !$user->isBlock;
+        return $user->isA('customer');
     }
 
     /**
      * Determine whether the user can update the model.
      *
      * @param  \App\Models\User  $user
-     * @param  \App\JsonApi\Proxies\CompanyPartner  $company
+     * @param  Company  $company
      * @return \Illuminate\Auth\Access\Response|bool
      */
-    public function update(User $user, Company $company)
+    public function update(User $user, Company  $company)
     {
-        return false;
-        // return $user->id == $company->admin_user_id || $user->isA('superadmin', "admin");
+        // и если есть право
+        // и если является админом
+        //dd($partner);
+        //exit;
+        return  $user->id == $company->admin_user_id && $user->isA('customer', 'customerAdmin');
     }
 
     /**
      * Determine whether the user can delete the model.
      *
      * @param  \App\Models\User  $user
-     * @param  \App\JsonApi\Proxies\CompanyPartner  $company
+     * @param  Company  $company
      * @return \Illuminate\Auth\Access\Response|bool
      */
-    public function delete(User $user, Company $company)
+    public function delete(User $user, Company  $company)
     {
         return false;
-        // return $user->id == $company->admin_user_id || $user->isA('superadmin', "admin");
+        //return  $user->id == $company->admin_user_id && $user->isA('customer', 'customerAdmin');
     }
 
     /**
      * Determine whether the user can restore the model.
      *
      * @param  \App\Models\User  $user
-     * @param  \App\JsonApi\Proxies\CompanyPartner  $company
+     * @param  Company  $company
      * @return \Illuminate\Auth\Access\Response|bool
      */
-    public function restore(User $user, Company $company)
+    public function restore(User $user, Company  $company)
     {
-        //return $user->id == $company->admin_user_id || $user->isA('superadmin', "admin");
         return false;
     }
 
@@ -89,10 +90,10 @@ class CompanyPolicy
      * Determine whether the user can permanently delete the model.
      *
      * @param  \App\Models\User  $user
-     * @param  \App\JsonApi\Proxies\CompanyPartner  $company
+     * @param  Company  $company
      * @return \Illuminate\Auth\Access\Response|bool
      */
-    public function forceDelete(User $user, Company $company)
+    public function forceDelete(User $user, Company  $company)
     {
         return false;
     }