+++ app/Policies/Customer/PartnerPolicy.php
@@ -1,93 +1,85 @@
 <?php
 
-namespace App\Policies\Partner;
+namespace App\Policies\Customer;
 
-use App\JsonApi\Proxies\PartnerUserPartner as PartnerUser;
+use App\JsonApi\Proxies\PartnerCustomer as Partner;
 use App\Models\User;
 use Illuminate\Auth\Access\HandlesAuthorization;
 
-class PartnerUserPolicy
+class PartnerPolicy
 {
     use HandlesAuthorization;
 
     /**
      * Determine whether the user can view any models.
      *
-     * @param  \App\Models\User $user
+     * @param  \App\Models\User  $user
      * @return \Illuminate\Auth\Access\Response|bool
      */
     public function viewAny(User $user)
     {
-        return $user->isA('partnerAdmin');
+        return true;
        // return $user->isA('partner');
     }
 
     /**
      * Determine whether the user can view the model.
      *
-     * @param  User $user
-     * @param  PartnerUser $partnerUser
+     * @param  \App\Models\User  $user
+     * @param  Partner $partner
      * @return \Illuminate\Auth\Access\Response|bool
      */
-    public function view(User $user, PartnerUser $partnerUser)
+    public function view(User $user, Partner $partner)
     {
-        return $user->isA('partnerAdmin');
+        return $user->isA('customer');
     }
 
     /**
      * Determine whether the user can create models.
      *
-     * @param  User $user
+     * @param  \App\Models\User  $user
      * @return \Illuminate\Auth\Access\Response|bool
      */
     public function create(User $user)
     {
-        /**
-         * проверка на то что можно создавать сотрудника с привязкой к этому партнеру
-         *
-         * получить partnerId
-         * получить partner
-         */
-
-        return $user->isA('partnerAdmin');
+        return false;
         // return !$user->isBlock;
     }
 
     /**
      * Determine whether the user can update the model.
      *
-     * @param  User $user
-     * @param  PartnerUser $partnerUser
+     * @param  \App\Models\User  $user
+     * @param  Partner $partner
      * @return \Illuminate\Auth\Access\Response|bool
      */
-    public function update(User $user, PartnerUser $partnerUser)
+    public function update(User $user, Partner $partner)
     {
-        return $user->isA('partnerAdmin');
+        return false;
         // return $user->id == $company->admin_user_id || $user->isA('superadmin', "admin");
     }
 
     /**
      * Determine whether the user can delete the model.
      *
-     * @param  User $user
-     * @param  PartnerUser $partnerUser
+     * @param  \App\Models\User  $user
+     * @param  Partner $partner
      * @return \Illuminate\Auth\Access\Response|bool
      */
-    public function delete(User $user, PartnerUser $partnerUser)
+    public function delete(User $user, Partner $partner)
     {
-        return $user->id != $partnerUser->is_admin && $user->isA('partnerAdmin');
-        //return false;
+        return false;
         // return $user->id == $company->admin_user_id || $user->isA('superadmin', "admin");
     }
 
     /**
      * Determine whether the user can restore the model.
      *
-     * @param  User $user
-     * @param  PartnerUser $partner
+     * @param  \App\Models\User  $user
+     * @param  Partner $partner
      * @return \Illuminate\Auth\Access\Response|bool
      */
-    public function restore(User $user, PartnerUser $partner)
+    public function restore(User $user, Partner $partner)
     {
         return false;
     }
@@ -95,11 +87,11 @@ class PartnerUserPolicy
     /**
      * Determine whether the user can permanently delete the model.
      *
-     * @param  User $user
-     * @param  PartnerUser $partnerUser
+     * @param  \App\Models\User  $user
+     * @param  \App\JsonApi\Proxies\CompanyPartner  $company
      * @return \Illuminate\Auth\Access\Response|bool
      */
-    public function forceDelete(User $user, PartnerUser $partnerUser)
+    public function forceDelete(User $user, Partner $partner)
     {
         return false;
     }