+++ app/Providers/AppServiceProvider.php
@@ -2,6 +2,8 @@
 
 namespace App\Providers;
 
+use DB;
+use Log;
 use Illuminate\Support\ServiceProvider;
 
 class AppServiceProvider extends ServiceProvider
@@ -23,6 +25,16 @@ class AppServiceProvider extends ServiceProvider
      */
     public function boot()
     {
-        //
+        // https://stackoverflow.com/a/57519919/12562591
+        /*
+        DB::listen(function($query) {
+            Log::info(
+                $query->sql,
+                $query->bindings,
+                //$query->time
+            );
+        });
+        */
+
     }
 }